+++ src/falconpy/_endpoint/__init__.py
@@ -55,6 +55,7 @@ from ._ioc import _ioc_endpoints
 from ._iocs import _iocs_endpoints
 from ._kubernetes_protection import _kubernetes_protection_endpoints
 from ._malquery import _malquery_endpoints
+from ._message_center import _message_center_endpoints
 from ._ml_exclusions import _ml_exclusions_endpoints
 from ._mssp import _mssp_endpoints
 from ._oauth2 import _oauth2_endpoints
@@ -101,6 +102,7 @@ api_endpoints.extend(_ioc_endpoints)
 api_endpoints.extend(_iocs_endpoints)
 api_endpoints.extend(_kubernetes_protection_endpoints)
 api_endpoints.extend(_malquery_endpoints)
+api_endpoints.extend(_message_center_endpoints)
 api_endpoints.extend(_ml_exclusions_endpoints)
 api_endpoints.extend(_mssp_endpoints)
 api_endpoints.extend(_oauth2_endpoints)